Rename app to Infantem Closes #17
