+++ .gitignore
@@ -39,3 +39,7 @@ Thumbs.db
 
 *.swp
 *.swo
+
+*.vim/
+*workspace/
+
