+++ src/main/java/com/isppG8/infantem/infantem/example/ExampleController.java
@@ -1,4 +1,4 @@
-package com.isppG8.friendsontour.friends_on_tour.example;
+package com.isppG8.infantem.infantem.example;
 
 import java.util.List;
