+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
-spring.application.name=friends-on-tour
-spring.datasource.url=jdbc:mysql://localhost:3306/friends_on_tour_db
+spring.application.name=infantem
+spring.datasource.url=jdbc:mysql://localhost:3306/infantem_db
 
 # Important: Currently the environments are hardcoded. 
 # This can be changed, but the env vars must be real system environment variables.
